vote-interface: move versions tests into vote_state_versions.rs
Move 3 tests that exercise VoteStateVersions APIs out of mod.rs and
into vote_state_versions.rs: default_vote_state_is_uninitialized,
is_correct_size_and_initialized, and version_conversion_bls_pubkey.

mod.rs now only retains tests for types defined there: CircBuf and
serde_compact_vote_state_update.
